requirements.txt
streamlit
python-dotenv
openai
pypdf
gradio
requests